+++ src/FiscalYears/FiscalYearDetails/FiscalYearGroups/FiscalYearGroups.js
@@ -1,27 +1,25 @@
-import React, {
+import PropTypes from 'prop-types';
+import {
   useCallback,
   useState,
   useEffect,
 } from 'react';
-import PropTypes from 'prop-types';
-import ReactRouterPropTypes from 'react-router-prop-types';
 import { withRouter } from 'react-router-dom';
+import ReactRouterPropTypes from 'react-router-prop-types';
 
 import { LIMIT_MAX } from '@folio/stripes-acq-components';
 import { stripesConnect } from '@folio/stripes/core';
 
 import { GROUPS_ROUTE } from '../../../common/const';
-import {
-  groupsResource,
-} from '../../../common/resources';
-import {
-  getGroupsWithTotals,
-} from '../../../common/utils';
+import { groupsResource } from '../../../common/resources';
+import { getGroupsWithTotals } from '../../../common/utils';
 import ConnectionListing from '../../../components/ConnectionListing';
+
+const DEFAULT_GROUP_SUMMARIES = [];
 
 const FiscalYearGroups = ({
   fiscalYear,
-  groupSummaries,
+  groupSummaries = DEFAULT_GROUP_SUMMARIES,
   history,
   mutator,
 }) => {
@@ -77,10 +75,6 @@ FiscalYearGroups.propTypes = {
   groupSummaries: PropTypes.arrayOf(PropTypes.object),
   history: ReactRouterPropTypes.history.isRequired,
   mutator: PropTypes.object.isRequired,
-};
-
-FiscalYearGroups.defaultProps = {
-  groupSummaries: [],
 };
 
 export default withRouter(stripesConnect(FiscalYearGroups));